Use maven to build the web services instead of ant
The need for dependency resolution in the web services pushed the module
to use maven instead of ant. But instead of using maven to build the whole
project, only use it to build the web services. Ant is still used to compile
and package the android module because I don't trust the maven android
plugins nor the ant maven interactivity plugins.